+++ src/sickle_ext/control.rs
@@ -15,7 +15,7 @@ fn collect_dangling_controlled(child: Entity, world: &World, dangling: &mut Vec<
     }
 
     // Collect dangling label.
-    if world.get::<ControlLabel>(child).is_some() {
+    if world.get::<ControlMember>(child).is_some() {
         dangling.push(child);
     }
 
@@ -31,22 +31,28 @@ fn collect_dangling_controlled(child: Entity, world: &World, dangling: &mut Vec<
 
 /// Instruction loadable for setting up the root entity of a multi-entity widget.
 ///
-/// Applies a [`ControlLabel`] instruction and inserts an internal `ControlMap` component to the entity.
+/// Applies a [`ControlMember`] instruction and inserts an internal `ControlMap` component to the entity.
 ///
-/// Children of the root node can be accessed through their [`ControlLabels`](ControlLabel) using
+/// Children of the root node can be accessed through their [`ControlMembers`](ControlMember) using
 /// [`ControlBuilderExt::edit_child`].
 ///
 /// It is recommended to apply this instruction before `Static`/`Responsive`/`Animated` instructions for optimal
 /// performance.
 #[derive(Reflect, Default, Clone, Debug, Deref, DerefMut, Eq, PartialEq)]
-pub struct ControlRoot(#[reflect(default)] pub SmolStr);
+pub struct ControlRoot
+{
+    /// Optional ID for the root. Use this if you want to use [`Responsive::respond_to`] or
+    /// [`Animated::respond_to`].
+    #[reflect(default)]
+    pub id: SmolStr,
+}
 
 impl ControlRoot
 {
     /// Makes a control root from a static string.
-    pub fn new(label: &'static str) -> Self
+    pub fn new(id: &'static str) -> Self
     {
-        Self(SmolStr::new_static(label))
+        Self { id: SmolStr::new_static(id) }
     }
 }
 
@@ -54,7 +60,7 @@ impl<T: Into<SmolStr>> From<T> for ControlRoot
 {
     fn from(string: T) -> Self
     {
-        Self(string.into())
+        Self { id: string.into() }
     }
 }
 
@@ -85,8 +91,8 @@ impl Instruction for ControlRoot
             if emut.contains::<Children>() {
                 // Look for the nearest ancestor with non-anonymous ControlMap and
                 // force it to re-apply its labels, in case this new root needs to steal some.
-                // - Note: we don't check for ControlLabel here since any pre-existing ControlLabel was likely
-                //   removed when the entity was switched to ControlRoot. When a ControlLabel is removed, the
+                // - Note: we don't check for ControlMember here since any pre-existing ControlMember was likely
+                //   removed when the entity was switched to ControlRoot. When a ControlMember is removed, the
                 //   entity will be removed from the associated ControlMap, which should ensure no stale attributes
                 //   related to this entity will linger in other maps.
                 let mut current = entity;
@@ -101,13 +107,13 @@ impl Instruction for ControlRoot
                     // Labels
                     // - Re-applying these forces the entities to re-register in the correct control maps.
                     for (label, label_entity) in labels {
-                        ControlLabel(label).apply(label_entity, world);
+                        ControlMember { id: label }.apply(label_entity, world);
                     }
 
                     break;
                 }
 
-                // Iterate children (stopping at control maps) to identify children with ControlLabel. Refresh
+                // Iterate children (stopping at control maps) to identify children with ControlMember. Refresh
                 // those nodes in case they have attributes that are 'dangling'.
                 let mut dangling = vec![];
                 for child in world.get::<Children>(entity).unwrap().iter() {
@@ -122,7 +128,7 @@ impl Instruction for ControlRoot
         }
 
         // Update control label.
-        ControlLabel(self.0).apply(entity, world);
+        ControlMember::from(self.id).apply(entity, world);
     }
 
     fn revert(entity: Entity, world: &mut World)
@@ -134,15 +140,16 @@ impl Instruction for ControlRoot
         }
 
         // Clean up the label.
-        ControlLabel::revert(entity, world);
+        ControlMember::revert(entity, world);
     }
 }
 
 //-------------------------------------------------------------------------------------------------------------------
 
-/// Instruction loadable that adds self as a component to indicate the entity is part of a widget.
+/// Instruction loadable that adds self as a component to indicate the entity is part of a multi-entity control
+/// group.
 ///
-/// Use this if you want values on the entity to respond to interactions on other parts of the widget.
+/// Use this if you want values on the entity to respond to interactions on other parts of the group.
 ///
 /// Values in a multi-entity widget can be controlled with the [`Static`], [`Responsive`], and [`Animated`]
 /// loadables.
@@ -150,45 +157,49 @@ impl Instruction for ControlRoot
 /// It is recommended to apply this instruction before `Static`/`Responsive`/`Animated` instructions for optimal
 /// performance.
 #[derive(Component, Reflect, Default, Clone, Debug, Deref, DerefMut, Eq, PartialEq)]
-pub struct ControlLabel(#[reflect(default)] pub SmolStr);
+pub struct ControlMember
+{
+    /// Optional ID for the control group member. Use this if you want to use [`Responsive::respond_to`] or
+    /// [`Animated::respond_to`].
+    #[reflect(default)]
+    pub id: SmolStr,
+}
 
-impl ControlLabel
+impl ControlMember
 {
     /// Makes a label from a static string.
     pub fn new(label: &'static str) -> Self
     {
-        Self(SmolStr::new_static(label))
+        Self { id: SmolStr::new_static(label) }
     }
 }
 
-impl<T: Into<SmolStr>> From<T> for ControlLabel
+impl<T: Into<SmolStr>> From<T> for ControlMember
 {
     fn from(string: T) -> Self
     {
-        Self(string.into())
+        Self { id: string.into() }
     }
 }
 
-impl Instruction for ControlLabel
+impl Instruction for ControlMember
 {
     fn apply(mut self, entity: Entity, world: &mut World)
     {
         let Ok(mut emut) = world.get_entity_mut(entity) else { return };
 
         // Anonymous label
-        if self.len() == 0 {
+        if self.id.len() == 0 {
             // TODO: smol_str v0.3 lets you do this without intermediate allocation
-            self = ControlLabel(SmolStr::from(
-                format!("_anon-{}v{}", entity.index(), entity.generation()),
-            ));
+            self.id = SmolStr::from(format!("_anon-{}v{}", entity.index(), entity.generation()));
         }
 
         // Insert or update control label.
-        let label_str = self.0.clone();
-        if let Some(mut existing_label) = emut.get_mut::<ControlLabel>() {
-            if *existing_label != self {
-                tracing::warn!("updating control label on {entity:?} from {existing_label:?} to {self:?}");
-                *existing_label = self;
+        let label_str = self.id.clone();
+        if let Some(mut existing_member) = emut.get_mut::<ControlMember>() {
+            if existing_member.id != self.id {
+                tracing::warn!("updating control label on {entity:?} from {existing_member:?} to {self:?}");
+                existing_member.id = self.id;
             }
         } else {
             emut.insert(self);
@@ -217,7 +228,7 @@ impl Instruction for ControlLabel
         }
 
         tracing::error!(
-            "error while inserting ControlLabel({label_str}) to {entity:?}, no ancestor with ControlMap \
+            "error while inserting ControlMember({label_str}) to {entity:?}, no ancestor with ControlMap \
             (see ControlRoot)",
         );
     }
@@ -253,7 +264,7 @@ impl Instruction for ControlLabel
 
         // Remove label component.
         let mut emut = world.entity_mut(entity);
-        emut.remove::<ControlLabel>();
+        emut.remove::<ControlMember>();
     }
 }
 
@@ -266,7 +277,7 @@ impl Plugin for ControlPlugin
     fn build(&self, app: &mut App)
     {
         app.register_instruction_type::<ControlRoot>()
-            .register_instruction_type::<ControlLabel>();
+            .register_instruction_type::<ControlMember>();
     }
 }